timers hooked up to data model

+++ XML/TileData.cpp
@@ -2,7 +2,7 @@
 
 TileData::TileData() : timerElapsed(0) {}
 
-TileData::TileData(const std::string& lbl, int timeElapsed) 
+TileData::TileData(const std::string& lbl, long long timeElapsed) 
     : label(lbl), timerElapsed(timeElapsed) {}
 
 void TileData::addSubTask(const std::string& subTask) {
