+++ XML/TileData.h
@@ -8,10 +8,10 @@ class TileData {
 public:
     std::string label; // Label of the tile
     std::vector<std::string> subTasks; // List of sub-tasks
-    int timerElapsed; // Time elapsed
+    long long timerElapsed;
 
     TileData();
-    TileData(const std::string& lbl, int timeElapsed);
+    TileData(const std::string& lbl, long long timeElapsed);
 
     void addSubTask(const std::string& subTask);
 };
